chore: split app.tsx components
